# swagger.yaml
openapi: 3.0.0
info:
  title: Valet NYC API
  version: 1.0.0
  description: API documentation for Valet NYC backend service
  contact:
    name: Valet NYC Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://valetnycdev.nfndev.com/api
    description: Dev server
  - url: https://valetnyc-api-uat.nfndev.com/api
    description: UAT server
  - url: https://valetnyc-api-production.nfndev.com/api
    description: Temp Prod server
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
    KeyTag:
      type: object
      properties:
        keyTagNumberold:
          type: integer
          minimum: 1
          maximum: 200
          description: Unique key tag number
        orderId:
          type: string
          description: Associated order ID
        vehicle:
          type: object
          properties:
            make:
              type: string
              description: Vehicle make/brand
            model:
              type: string
              description: Vehicle model
            color:
              type: string
              description: Vehicle color
            licensePlate:
              type: string
              description: Vehicle license plate
            keyTagNumber:
              type: string
              description: Key tag number
        parkingLocation:
          type: string
          description: Parking location address
    KeyTransfer:
      type: object
      properties:
        _id:
          type: string
          description: Transfer ID
        senderValet:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            phone:
              type: string
        receiverValet:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            phone:
              type: string
        keys:
          type: array
          items:
            $ref: '#/components/schemas/KeyTag'
        status:
          type: string
          enum: [pending_acceptance, accepted, rejected, cancelled]
          description: Transfer status
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
        rejectedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
        cancelledReason:
          type: string
        receiverConfirmedKeyTags:
          type: array
          items:
            type: integer
    KeyTransferAudit:
      type: object
      properties:
        _id:
          type: string
        transferId:
          type: string
        senderValet:
          type: object
          properties:
            _id:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            phone:
              type: string
        receiverValet:
          type: object
          properties:
            _id:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            phone:
              type: string
        keyCount:
          type: integer
        keyTags:
          type: array
          items:
            type: integer
        action:
          type: string
          enum: [initiated, accepted, rejected, cancelled]
        status:
          type: string
          enum: [pending_acceptance, accepted, rejected, cancelled]
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
        ipAddress:
          type: string
        userAgent:
          type: string
    Valet:
      type: object
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        profileImage:
          type: string
...
paths:
...
  /notification/send:
    post:
      summary: Send notification to user
      tags: [Notification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                title:
                  type: string
                body:
                  type: string
                data:
                  type: object
              required: [userId, title, body]
      responses:
        '200':
          description: Notification sent successfully
        '503':
          description: Firebase not configured
  /notification/notify-closest-valets:
    post:
      summary: Notify closest valets to pickup location
      tags: [Notification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pickupLocation:
                  type: object
                orderId:
                  type: string
                radius:
                  type: number
                  description: Radius in kilometers
              required: [pickupLocation, orderId]
      responses:
        '200':
          description: Valets notified successfully
        '503':
          description: Firebase not configured
  /notification/send-push:
    post:
      summary: Send push notification to a single user
      tags: [Notification]
      description: Send a push notification to a specific user using their Firebase UID. Retrieves FCM token from Firestore and sends notification via Firebase Cloud Messaging.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firebaseUid:
                  type: string
                  description: Firebase UID of the recipient
                title:
                  type: string
                  description: Notification title
                body:
                  type: string
                  description: Notification body/message
                data:
                  type: object
                  description: Optional custom data payload to include with notification
                  example:
                    type: key_transfer_request
                    transferId: "507f1f77bcf86cd799439011"
              required: [firebaseUid, title, body]
      responses:
        '200':
          description: Push notification sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  response:
                    type: object
                    description: Firebase messaging response
        '400':
          description: Invalid request or user does not have valid FCM token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Server error while sending notification
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
...
  /key-transfer/my-keys:
    get:
      summary: Get all available keys for transfer
      tags: [Key Transfer]
      description: Retrieves list of keys currently assigned to the valet that are available for transfer
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
          description: Valet user ID
      responses:
        '200':
          description: List of available keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyTag'
                  count:
                    type: integer
        '401':
          description: User ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only valets can access key transfer features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /key-transfer/outgoing:
    get:
      summary: Get outgoing transfer requests
      tags: [Key Transfer]
      description: Retrieves all outgoing key transfer requests initiated by the valet
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
          description: Valet user ID
      responses:
        '200':
          description: List of outgoing transfers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        receiverValet:
                          type: object
                          properties:
                            _id:
                              type: string
                            name:
                              type: string
                            phone:
                              type: string
                        keys:
                          type: array
                          items:
                            $ref: '#/components/schemas/KeyTag'
                        status:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        canCancel:
                          type: boolean
                  count:
                    type: integer
  /key-transfer/incoming:
    get:
      summary: Get incoming transfer requests
      tags: [Key Transfer]
      description: Retrieves all incoming key transfer requests sent to the valet
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
          description: Valet user ID
      responses:
        '200':
          description: List of incoming transfers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        senderValet:
                          type: object
                          properties:
                            _id:
                              type: string
                            name:
                              type: string
                            phone:
                              type: string
                        keys:
                          type: array
                          items:
                            $ref: '#/components/schemas/KeyTag'
                        status:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        canRespond:
                          type: boolean
                  count:
                    type: integer
  /key-transfer/active-valets:
    get:
      summary: Get list of active valets
      tags: [Key Transfer]
      description: Retrieves list of active valets available to receive key transfers
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
          description: Current valet user ID
      responses:
        '200':
          description: List of active valets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Valet'
                  count:
                    type: integer
  /key-transfer/audit-logs:
    get:
      summary: Get key transfer audit logs
      tags: [Key Transfer]
      description: Retrieves immutable audit trail of all key transfer activities
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
          description: Current valet user ID
        - in: query
          name: transferId
          schema:
            type: string
          description: Filter by transfer ID
        - in: query
          name: senderValetId
          schema:
            type: string
          description: Filter by sender valet ID
        - in: query
          name: receiverValetId
          schema:
            type: string
          description: Filter by receiver valet ID
        - in: query
          name: action
          schema:
            type: string
            enum: [initiated, accepted, rejected, cancelled]
          description: Filter by action type
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Number of records to return
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
          description: Number of records to skip
      responses:
        '200':
          description: Audit logs with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyTransferAudit'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
                      skip:
                        type: integer
                      hasMore:
                        type: boolean
  /key-transfer/initiate:
    post:
      summary: Initiate a key transfer request
      tags: [Key Transfer]
      description: Creates a new key transfer request to send keys to another valet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: Sender valet ID
                receiverValetId:
                  type: string
                  description: Receiver valet ID
                keyOrderIds:
                  type: array
                  items:
                    type: string
                  description: Array of order IDs for keys to transfer
              required: [userId, receiverValetId, keyOrderIds]
      responses:
        '201':
          description: Transfer request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/KeyTransfer'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /key-transfer/accept:
    post:
      summary: Accept a key transfer request
      tags: [Key Transfer]
      description: Accepts an incoming key transfer request and confirms receipt of all key tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: Receiver valet ID
                transferId:
                  type: string
                  description: Transfer request ID
                confirmedKeyTags:
                  type: array
                  items:
                    type: integer
                  description: Array of confirmed key tag numbers
              required: [userId, transferId, confirmedKeyTags]
      responses:
        '200':
          description: Transfer accepted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/KeyTransfer'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized to accept this transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transfer request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /key-transfer/reject:
    post:
      summary: Reject a key transfer request
      tags: [Key Transfer]
      description: Rejects an incoming key transfer request with optional reason
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: Receiver valet ID
                transferId:
                  type: string
                  description: Transfer request ID
                rejectionReason:
                  type: string
                  description: Optional reason for rejection
              required: [userId, transferId]
      responses:
        '200':
          description: Transfer rejected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/KeyTransfer'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized to reject this transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transfer request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /key-transfer/cancel:
    post:
      summary: Cancel a key transfer request
      tags: [Key Transfer]
      description: Cancels an outgoing key transfer request (only if status is pending_acceptance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: Sender valet ID
                transferId:
                  type: string
                  description: Transfer request ID
                cancelReason:
                  type: string
                  description: Optional reason for cancellation
              required: [userId, transferId]
      responses:
        '200':
          description: Transfer cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/KeyTransfer'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized to cancel this transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transfer request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...